widgets-dock | use desktop as window type
a change in fedora 39 causes the sidebar to come in focus when the mouse
is hovered on top of it.  this causes the sidebar to come on top
of the sidebar panel, effectively hiding it.

removing the use of the panel window type, fixes the problem at the
expense of losing the panel effect.